+++ astropy/units/function/logarithmic.py
@@ -4,9 +4,8 @@ from __future__ import (absolute_import, unicode_literals,
                         division, print_function)
 import numpy as np
 
-from .. import (CompositeUnit, Unit, UnitsError, dimensionless_unscaled,
-                si, astrophys)
-from ...constants import L_bol0
+from .. import CompositeUnit, UnitsError, dimensionless_unscaled
+from . import magnitude_zero_points as mag0
 from .core import FunctionUnitBase, FunctionQuantity
 from .units import dex, dB, mag
 
@@ -120,6 +119,11 @@ class MagUnit(LogUnit):
         By default, this is ``mag``, but this allows one to use an equivalent
         unit such as ``2 mag``.
     """
+    def __init__(self, *args, **kwargs):
+        # Ensure we recognize magnitude zero points here.
+        with mag0.enable():
+            super(MagUnit, self).__init__(*args, **kwargs)
+
     @property
     def _default_function_unit(self):
         return mag
@@ -317,30 +321,16 @@ dB._function_unit_class = DecibelUnit
 mag._function_unit_class = MagUnit
 
 
-AB0 = Unit('AB', 10.**(-0.4*48.6) * 1.e-3 * si.W / si.m**2 / si.Hz,
-           doc="AB magnitude zero flux density.")
-
-ST0 = Unit('ST', 10.**(-0.4*21.1) * 1.e-3 * si.W / si.m**2 / si.AA,
-           doc="ST magnitude zero flux density.")
-
-Bol0 = Unit('Bol', L_bol0, doc="Luminosity corresponding to "
-            "absolute bolometric magnitude zero")
-
-bol0 = Unit('bol', L_bol0 / (4 * np.pi * (10.*astrophys.pc)**2),
-            doc="Irradiance corresponding to apparent bolometric magnitude "
-            "zero")
-
-
-STmag = MagUnit(ST0)
+STmag = MagUnit(mag0.ST)
 STmag.__doc__ = "ST magnitude: STmag=-21.1 corresponds to 1 erg/s/cm2/A"
 
-ABmag = MagUnit(AB0)
+ABmag = MagUnit(mag0.AB)
 ABmag.__doc__ = "AB magnitude: ABmag=-48.6 corresponds to 1 erg/s/cm2/Hz"
 
-M_bol = MagUnit(Bol0)
+M_bol = MagUnit(mag0.Bol)
 M_bol.__doc__ = ("Absolute bolometric magnitude: M_bol=0 corresponds to "
-                 "L_bol0={0}".format(Bol0.si))
+                 "L_bol0={0}".format(mag0.Bol.si))
 
-m_bol = MagUnit(bol0)
+m_bol = MagUnit(mag0.bol)
 m_bol.__doc__ = ("Apparent bolometric magnitude: m_bol=0 corresponds to "
-                 "f_bol0={0}".format(bol0.si))
+                 "f_bol0={0}".format(mag0.bol.si))