+++ CHANGES.rst
@@ -40,6 +40,11 @@ New Features
 
   - Added ``pixel_scale`` and ``plate_scale`` equivalencies. [#4987]
 
+  - Magnitude zero points used to define ``STmag``, ``ABmag``, ``M_bol`` and
+    ``m_bol`` are now collected in ``astropy.units.magnitude_zero_points``.
+    They are not enabled as regular units by default, but can be included
+    using ``astropy.units.magnitude_zero_points.enable()``. [#5030]
+
 - ``astropy.utils``
 
 - ``astropy.vo``